+++ src/main/java/com/nisaelek/springtoDoApp/repos/TodoRepository.java
@@ -3,5 +3,5 @@ package com.nisaelek.springtoDoApp.repos;
 import com.nisaelek.springtoDoApp.entities.Todo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TodoRepository extends JpaRepository<Todo,Integer> {
+public interface TodoRepository extends JpaRepository<Todo,Long> {
 }
