+++ src/main/java/com/nisaelek/springtoDoApp/repos/UserRepository.java
@@ -3,5 +3,6 @@ package com.nisaelek.springtoDoApp.repos;
 import com.nisaelek.springtoDoApp.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository <User,Integer> {
+public interface UserRepository extends JpaRepository <User,Long> {
+
 }
